+++ src/components/float/SnowflakeHandler.tsx
@@ -1,6 +1,6 @@
 import { IconSnow } from '@components/icon';
 import { CircleIconWrapper } from '@components/icon/wrapper';
-import { useSnowflakeStore } from '@store/configStore';
+import { useSnowflakeStore } from '@store/config';
 import React from 'react';
 
 export default function SnowflakeHandler() {
