+++ src/hooks/useTheme.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
-import { CONFIG_THEME_KEY } from '@src/constants';
+import { CONFIG_THEME_KEY, Themes } from '@src/constants';
 import { useLocalStorage } from './useLocalStorage';
-import { Themes, useThemeStore } from '@store/configStore';
+import { useThemeStore } from '@store/config';
 
 const useTheme = () => {
   const { setConfig, getConfig } = useLocalStorage();
