+++ src/store/context.ts
@@ -1,12 +1,12 @@
 import React from 'react';
 import { MultiSelect, NotionNode } from '@types';
 
-export interface INotionContext {
+export interface NotionContextProps {
   nodes: NotionNode[];
   everyPostsTags?: string[];
   everyPostsSeries?: MultiSelect;
 }
-export const NotionContext = React.createContext<INotionContext>({
+export const NotionContext = React.createContext<NotionContextProps>({
   nodes: [],
   everyPostsTags: [],
   everyPostsSeries: [],
