+++ src/store/index.ts
@@ -1,0 +1,2 @@
+export * from './context';
+export * from './config';
